# services/order-demo/internal/adapters/inbound/http/generated/order_api.gen.go
//go:build go1.22

// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package generated

import (
	"context"
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	OperationalReadTokenScopes operationalReadTokenContextKey = "OperationalReadToken.Scopes"
	RemediationTokenScopes     remediationTokenContextKey     = "RemediationToken.Scopes"
)

// Defines values for ErrorCode.
const (
	IdempotencyConflict      ErrorCode = "idempotency_conflict"
	InternalError            ErrorCode = "internal_error"
	InvalidArgument          ErrorCode = "invalid_argument"
	OperationNotFound        ErrorCode = "operation_not_found"
	QueueFull                ErrorCode = "queue_full"
	ResourceNotFound         ErrorCode = "resource_not_found"
	TargetPreconditionFailed ErrorCode = "target_precondition_failed"
	Unauthenticated          ErrorCode = "unauthenticated"
)

// Valid indicates whether the value is a known member of the ErrorCode enum.
func (e ErrorCode) Valid() bool {
	switch e {
	case IdempotencyConflict:
		return true
	case InternalError:
		return true
	case InvalidArgument:
		return true
	case OperationNotFound:
		return true
	case QueueFull:
		return true
	case ResourceNotFound:
		return true
	case TargetPreconditionFailed:
		return true
	case Unauthenticated:
		return true
	default:
		return false
	}
}

// Defines values for OrderFailureReason.
const (
	DependencyUnavailable OrderFailureReason = "dependency_unavailable"
	RetryExhausted        OrderFailureReason = "retry_exhausted"
)

// Valid indicates whether the value is a known member of the OrderFailureReason enum.
func (e OrderFailureReason) Valid() bool {
	switch e {
	case DependencyUnavailable:
		return true
	case RetryExhausted:
		return true
	default:
		return false
	}
}

// Defines values for OrderStatus.
const (
	OrderStatusCompleted  OrderStatus = "completed"
	OrderStatusFailed     OrderStatus = "failed"
	OrderStatusProcessing OrderStatus = "processing"
	OrderStatusQueued     OrderStatus = "queued"
)

// Valid indicates whether the value is a known member of the OrderStatus enum.
func (e OrderStatus) Valid() bool {
	switch e {
	case OrderStatusCompleted:
		return true
	case OrderStatusFailed:
		return true
	case OrderStatusProcessing:
		return true
	case OrderStatusQueued:
		return true
	default:
		return false
	}
}

// Defines values for ProbeResultResult.
const (
	ProbeResultResultCompleted ProbeResultResult = "completed"
	ProbeResultResultFailed    ProbeResultResult = "failed"
	ProbeResultResultTimedOut  ProbeResultResult = "timed_out"
)

// Valid indicates whether the value is a known member of the ProbeResultResult enum.
func (e ProbeResultResult) Valid() bool {
	switch e {
	case ProbeResultResultCompleted:
		return true
	case ProbeResultResultFailed:
		return true
	case ProbeResultResultTimedOut:
		return true
	default:
		return false
	}
}

// Defines values for RuntimeSnapshotSupervisorStatus.
const (
	Degraded RuntimeSnapshotSupervisorStatus = "degraded"
	Running  RuntimeSnapshotSupervisorStatus = "running"
)

// Valid indicates whether the value is a known member of the RuntimeSnapshotSupervisorStatus enum.
func (e RuntimeSnapshotSupervisorStatus) Valid() bool {
	switch e {
	case Degraded:
		return true
	case Running:
		return true
	default:
		return false
	}
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	Quantity int    `json:"quantity"`
	Sku      string `json:"sku"`
}

// Error defines model for Error.
type Error struct {
	Code      ErrorCode `json:"code"`
	Message   string    `json:"message"`
	Retryable bool      `json:"retryable"`
}

// ErrorCode defines model for Error.Code.
type ErrorCode string

// OperationReceipt defines model for OperationReceipt.
type OperationReceipt struct {
	AfterConcurrency  interface{} `json:"afterConcurrency"`
	AfterVersion      int         `json:"afterVersion"`
	ApprovalId        string      `json:"approvalId"`
	BeforeConcurrency interface{} `json:"beforeConcurrency"`
	BeforeVersion     int         `json:"beforeVersion"`
	ExecutedAt        time.Time   `json:"executedAt"`
	InstanceEpoch     string      `json:"instanceEpoch"`
	IntentDigest      string      `json:"intentDigest"`
	OperationId       string      `json:"operationId"`
}

// Order defines model for Order.
type Order struct {
	CreatedAt     time.Time           `json:"createdAt"`
	FailureReason *OrderFailureReason `json:"failureReason"`
	Id            string              `json:"id"`
	Status        OrderStatus         `json:"status"`
	UpdatedAt     time.Time           `json:"updatedAt"`
}

// OrderFailureReason defines model for Order.FailureReason.
type OrderFailureReason string

// OrderStatus defines model for OrderStatus.
type OrderStatus string

// ProbeResult defines model for ProbeResult.
type ProbeResult struct {
	CompletedAt *time.Time        `json:"completedAt"`
	DurationMs  int               `json:"durationMs"`
	OrderId     string            `json:"orderId"`
	ProbeId     string            `json:"probeId"`
	Result      ProbeResultResult `json:"result"`
}

// ProbeResultResult defines model for ProbeResult.Result.
type ProbeResultResult string

// QueueSnapshot defines model for QueueSnapshot.
type QueueSnapshot struct {
	Capacity         interface{} `json:"capacity"`
	Depth            int         `json:"depth"`
	ObservedAt       time.Time   `json:"observedAt"`
	OldestAgeSeconds float32     `json:"oldestAgeSeconds"`
}

// RuntimeSnapshot defines model for RuntimeSnapshot.
type RuntimeSnapshot struct {
	InstanceEpoch    string                          `json:"instanceEpoch"`
	ServiceRef       interface{}                     `json:"serviceRef"`
	StartedAt        time.Time                       `json:"startedAt"`
	SupervisorStatus RuntimeSnapshotSupervisorStatus `json:"supervisorStatus"`
}

// RuntimeSnapshotSupervisorStatus defines model for RuntimeSnapshot.SupervisorStatus.
type RuntimeSnapshotSupervisorStatus string

// UpdateWorkerConfigRequest defines model for UpdateWorkerConfigRequest.
type UpdateWorkerConfigRequest struct {
	ApprovalId          string      `json:"approvalId"`
	ExpectedConcurrency interface{} `json:"expectedConcurrency"`
	ExpectedVersion     int         `json:"expectedVersion"`
	InstanceEpoch       string      `json:"instanceEpoch"`
	IntentDigest        string      `json:"intentDigest"`
	NewConcurrency      interface{} `json:"newConcurrency"`
	OperationId         string      `json:"operationId"`
}

// WorkerConfig defines model for WorkerConfig.
type WorkerConfig struct {
	ActiveWorkers         int         `json:"activeWorkers"`
	ConfiguredConcurrency int         `json:"configuredConcurrency"`
	EffectiveConcurrency  int         `json:"effectiveConcurrency"`
	InflightOrders        int         `json:"inflightOrders"`
	InstanceEpoch         string      `json:"instanceEpoch"`
	ObservedAt            time.Time   `json:"observedAt"`
	ServiceRef            interface{} `json:"serviceRef"`
	Version               int         `json:"version"`
}

// WorkerPolicy defines model for WorkerPolicy.
type WorkerPolicy struct {
	Digest              string      `json:"digest"`
	ExpectedConcurrency interface{} `json:"expectedConcurrency"`
	MaxConcurrency      interface{} `json:"maxConcurrency"`
	MinConcurrency      interface{} `json:"minConcurrency"`
	ServiceRef          interface{} `json:"serviceRef"`
	Version             interface{} `json:"version"`
}

// IdempotencyKey defines model for IdempotencyKey.
type IdempotencyKey = string

// OperationId defines model for OperationId.
type OperationId = string

// OrderId defines model for OrderId.
type OrderId = string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = Error

// operationalReadTokenContextKey is the context key for OperationalReadToken security scheme
type operationalReadTokenContextKey string

// remediationTokenContextKey is the context key for RemediationToken security scheme
type remediationTokenContextKey string

// CreateOrderParams defines parameters for CreateOrder.
type CreateOrderParams struct {
	IdempotencyKey IdempotencyKey `json:"Idempotency-Key"`
}

// ListRecentOrdersParams defines parameters for ListRecentOrders.
type ListRecentOrdersParams struct {
	Status *OrderStatus `form:"status,omitempty" json:"status,omitempty"`
	Limit  *int         `form:"limit,omitempty" json:"limit,omitempty"`
}

// RunOrderProcessingProbeJSONBody defines parameters for RunOrderProcessingProbe.
type RunOrderProcessingProbeJSONBody struct {
	ProbeId string `json:"probeId"`
}

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// UpdateWorkerConfigJSONRequestBody defines body for UpdateWorkerConfig for application/json ContentType.
type UpdateWorkerConfigJSONRequestBody = UpdateWorkerConfigRequest

// RunOrderProcessingProbeJSONRequestBody defines body for RunOrderProcessingProbe for application/json ContentType.
type RunOrderProcessingProbeJSONRequestBody RunOrderProcessingProbeJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Accept an order into the bounded in-memory queue
	// (POST /api/v1/orders)
	CreateOrder(w http.ResponseWriter, r *http.Request, params CreateOrderParams)
	// Read a bounded order status
	// (GET /api/v1/orders/{orderId})
	GetOrder(w http.ResponseWriter, r *http.Request, orderId OrderId)
	// Read versioned worker configuration
	// (GET /ops/v1/config/worker)
	GetWorkerConfig(w http.ResponseWriter, r *http.Request)
	// Execute the one typed compare-and-swap worker remediation
	// (PUT /ops/v1/config/worker)
	UpdateWorkerConfig(w http.ResponseWriter, r *http.Request)
	// Reconcile a typed remediation operation
	// (GET /ops/v1/operations/{operationId})
	GetOperation(w http.ResponseWriter, r *http.Request, operationId OperationId)
	// Read bounded redacted recent outcomes
	// (GET /ops/v1/orders/recent)
	ListRecentOrders(w http.ResponseWriter, r *http.Request, params ListRecentOrdersParams)
	// Read the immutable healthy worker policy
	// (GET /ops/v1/policies/worker)
	GetWorkerPolicy(w http.ResponseWriter, r *http.Request)
	// Run a fixed bounded business probe
	// (POST /ops/v1/probes/order-processing)
	RunOrderProcessingProbe(w http.ResponseWriter, r *http.Request)
	// Read bounded queue health
	// (GET /ops/v1/queue)
	GetQueue(w http.ResponseWriter, r *http.Request)
	// Read service instance identity and worker supervisor state
	// (GET /ops/v1/runtime)
	GetRuntime(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateOrder operation middleware
func (siw *ServerInterfaceWrapper) CreateOrder(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateOrderParams

	headers := r.Header

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey IdempotencyKey
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err := fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateOrder(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOrder operation middleware
func (siw *ServerInterfaceWrapper) GetOrder(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "orderId" -------------
	var orderId OrderId

	err = runtime.BindStyledParameterWithOptions("simple", "orderId", r.PathValue("orderId"), &orderId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "orderId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOrder(w, r, orderId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWorkerConfig operation middleware
func (siw *ServerInterfaceWrapper) GetWorkerConfig(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWorkerConfig(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateWorkerConfig operation middleware
func (siw *ServerInterfaceWrapper) UpdateWorkerConfig(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, RemediationTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateWorkerConfig(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOperation operation middleware
func (siw *ServerInterfaceWrapper) GetOperation(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "operationId" -------------
	var operationId OperationId

	err = runtime.BindStyledParameterWithOptions("simple", "operationId", r.PathValue("operationId"), &operationId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "operationId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOperation(w, r, operationId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListRecentOrders operation middleware
func (siw *ServerInterfaceWrapper) ListRecentOrders(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ListRecentOrdersParams

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", r.URL.Query(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "status"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListRecentOrders(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWorkerPolicy operation middleware
func (siw *ServerInterfaceWrapper) GetWorkerPolicy(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWorkerPolicy(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RunOrderProcessingProbe operation middleware
func (siw *ServerInterfaceWrapper) RunOrderProcessingProbe(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RunOrderProcessingProbe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetQueue operation middleware
func (siw *ServerInterfaceWrapper) GetQueue(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetQueue(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRuntime operation middleware
func (siw *ServerInterfaceWrapper) GetRuntime(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, OperationalReadTokenScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRuntime(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/orders", wrapper.CreateOrder)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/orders/{orderId}", wrapper.GetOrder)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/config/worker", wrapper.GetWorkerConfig)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/ops/v1/config/worker", wrapper.UpdateWorkerConfig)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/operations/{operationId}", wrapper.GetOperation)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/orders/recent", wrapper.ListRecentOrders)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/policies/worker", wrapper.GetWorkerPolicy)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/ops/v1/probes/order-processing", wrapper.RunOrderProcessingProbe)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/queue", wrapper.GetQueue)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/ops/v1/runtime", wrapper.GetRuntime)

	return m
}